requirements.txt
aiodns==1.1.1
aiohttp==3.6.2
asn1crypto==1.1.0
async-timeout==3.0.1
attrs==19.3.0
ccxt==1.18.1279
certifi==2019.9.11
cffi==1.13.0
chardet==3.0.4
cryptography==2.7
Django==2.2.6
django-filter==2.2.0
django-redis-cache==2.1.0
djangorestframework==3.10.3
djangorestframework-stubs==1.0.0
idna==2.8
mpmath==1.1.0
multidict==4.5.2
mysqlclient==1.4.4
parse==1.12.1
pika==1.1.0
pycares==3.0.0
pycparser==2.19
PyJWT==1.7.1
python-dotenv==0.10.3
pytz==2019.3
redis==3.3.11
requests==2.22.0
requests-toolbelt==0.9.1
sentry-sdk==0.12.3
six==1.12.0
sqlparse==0.3.0
sympy==1.4
urllib3==1.25.6
yarl==1.1.0
